requirements.txt
Django>=4.2
psycopg2-binary>=2.9
cryptography>=41.0
python-dotenv>=1.0
Pillow>=10.0
gunicorn>=20.1
